Move Reminder.java and ReminderTest.java into project package
