+++ src/main/java/seedu/momentum/model/util/SampleDataUtil.java
@@ -12,7 +12,7 @@ import seedu.momentum.model.project.Deadline;
 import seedu.momentum.model.project.Description;
 import seedu.momentum.model.project.Name;
 import seedu.momentum.model.project.Project;
-import seedu.momentum.model.reminder.Reminder;
+import seedu.momentum.model.project.Reminder;
 import seedu.momentum.model.tag.Tag;
 import seedu.momentum.model.timer.UniqueDurationList;
 import seedu.momentum.model.timer.WorkDuration;
