+++ src/test/java/seedu/momentum/testutil/EditTrackedItemDescriptorBuilder.java
@@ -10,8 +10,8 @@ import seedu.momentum.model.project.CompletionStatus;
 import seedu.momentum.model.project.Deadline;
 import seedu.momentum.model.project.Description;
 import seedu.momentum.model.project.Name;
+import seedu.momentum.model.project.Reminder;
 import seedu.momentum.model.project.TrackedItem;
-import seedu.momentum.model.reminder.Reminder;
 import seedu.momentum.model.tag.Tag;
 
 /**
